+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.4.7
 - vs 18.5.2
+- swallow IOException in WriteTo to prevent JNI crash on android
 
 ## 2.4.6
 - vs 18.5.1